cleanup (#320)
