+++ app/features/pwa/use-should-show-pwa-prompt.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import useShouldShowPrompt from '~/hooks/use-should-show-prompt';
-import useUserAgent, { type Browser } from '~/hooks/use-user-agent';
+import useUserAgent from '~/hooks/use-user-agent';
 
 const key = 'showPwaPrompt';
 const TEMPORARY_DISMISS_DURATION_MS = 3 * 60 * 60 * 1000; // 3 hours
@@ -8,8 +8,6 @@ const TEMPORARY_DISMISS_DURATION_MS = 3 * 60 * 60 * 1000; // 3 hours
 type Return = {
   /** True if the PWA prompt should be shown based on the user's browser, device, and if the app is already installed */
   shouldShowPwaPrompt: boolean;
-  /** The user's browser parsed from the user agent string */
-  browser: Browser;
   /** Function to permanently dismiss the PWA prompt */
   dontShowAgain: () => void;
   /** Function to temporarily dismiss the PWA prompt */
@@ -20,7 +18,7 @@ type Return = {
 export default function useShouldShowPwaPrompt(): Return {
   const { shouldShow, handleDontShowAgain, handleDismissTemporarily } =
     useShouldShowPrompt(key);
-  const { isMobile, browser } = useUserAgent();
+  const { isMobile } = useUserAgent();
   const [isStandalone, setIsStandalone] = useState(false);
 
   useEffect(() => {
@@ -33,7 +31,6 @@ export default function useShouldShowPwaPrompt(): Return {
   const shouldShowPwaPrompt = shouldShow && isMobile && !isStandalone;
 
   return {
-    browser,
     shouldShowPwaPrompt,
     dontShowAgain: handleDontShowAgain,
     dismissTemporarily: () =>
